pyproject.toml
[project]
name = "mlops-pipeline-using-dvc-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.1",
    "scikit-learn>=1.4.0",
    "pyyaml>=6.0",
    "dvc>=3.0.0",
]
